MonsterChase#10 - Monster HSM and patrolling basic Behavior for monsters     Changes:     - behaviors.cpp/h: RandomDirection and PerpendicularDirection routing     behaviors     - monster.cpp/h: monsterFactory added, MonsterModel extended with     target_direction and type and position changed from int to float,     collisions with walls implementation     - monsterchase.cpp/h: small refactoring for adding multiple monstersm     - monstersm.cpp/h: HSM behaviors factory and smoothing direction changes     - monsterviews.cpp/h: added MonsterTriangularShape     - player.h: position changed from int to float into player model

+++ gtest/tst_collisions.h
@@ -32,8 +32,8 @@ public:
 
 TEST_F(Collisoins, debouncing)
 {
-    int start_x = 185;
-    int start_y = 200;
+    float start_x = 185;
+    float start_y = 200;
 
     model = new PlayerModel {
             DEF_ENERGY, //energy
